feat: refactor into previous approach
