+++ src/module.ts
@@ -1,13 +1,18 @@
-import { resolve } from 'path'
+import { resolve, dirname } from 'path'
 import { fileURLToPath } from 'url'
 import { defineNuxtModule, addPlugin } from '@nuxt/kit'
 import defu from 'defu'
 
+type AnalyticsProvider = 'log' | 'vercel' | 'ga'
 
-interface ModuleOptions {
-  provider: 'log' | 'vercel' | 'ga';
+export interface ModuleOptions {
+  provider: AnalyticsProvider;
   debug: boolean;
   disabled: boolean;
+  options?: {
+    googleAnalyticsId: string;
+    vercelAnalyticsDsn: string;
+  }
 }
 
 declare module '@nuxt/schema' {
@@ -15,8 +20,6 @@ declare module '@nuxt/schema' {
     webVitals: ModuleOptions
   }
 }
-
-// export * from './types'
 
 export default defineNuxtModule<ModuleOptions>({
   meta: {
@@ -29,27 +32,43 @@ export default defineNuxtModule<ModuleOptions>({
     disabled: false
   },
   setup (options, nuxt) {
-    let defaultProvider: string;
+    if (options.disabled) {
+      return
+    }
+
+    let defaultProvider: AnalyticsProvider;
 
     if (!options.provider || (options.provider !== 'ga' && options.provider !== 'vercel')) {
       defaultProvider = 'log'
       console.warn('`[@nuxtjs/web-vitals]` No Analytics providers selected. Using built in `log`')
     }
 
-    if (options.disabled) {
-      return
+    if (options.provider === 'ga' && !options.options.googleAnalyticsId) {
+      throw new Error('`[@nuxtjs/web-vitals]` Missing Google Analytics ID in module configuration')
+    }
+
+    if (options.provider === 'vercel' && !options.options.vercelAnalyticsDsn) {
+      throw new Error('`[@nuxtjs/web-vitals]` Missing Vercel Analytics DSN in module configuration')
     }
 
     nuxt.options.publicRuntimeConfig.webVitals = defu(nuxt.options.publicRuntimeConfig.webVitals, {
       provider: defaultProvider || options.provider,
       debug: options.debug,
       disabled: options.disabled,
+      options: options.options
     })
 
     nuxt.options.alias.ufo = 'ufo/dist/index.mjs'
 
+
     const runtimeDir = fileURLToPath(new URL('./runtime', import.meta.url))
     nuxt.options.build.transpile.push(runtimeDir)
+
+    const providerRuntime = require.resolve(`./runtime/providers/${options.provider}`)
+
+    nuxt.options.alias['~vitals-provider'] = providerRuntime
+    nuxt.options.build.transpile.push(dirname(providerRuntime))
+
     addPlugin(resolve(runtimeDir, 'plugin.client'))
 
     console.log('`[@nuxtjs/web-vitals]` Module loaded correctly 🚀')
